+++ src/c_linalg_lu.c
@@ -38,7 +38,7 @@ c_linalg_lapack_dgetrf (c_matrix *a, c_vector_int **p)
 	if (p) *p = _p;
 	else if (!c_vector_int_is_empty (_p)) c_vector_int_free (_p);
 
-	return (int) info;
+	return info;
 }
 
 int
@@ -61,7 +61,7 @@ c_linalg_lapack_dgetrs (char trans, c_matrix *lu, c_matrix *b, c_vector_int *p)
 	ldb = (int) b->lda;
 	dgetrs_ (&trans, &n, &nrhs, lu->data, &lda, p->data, b->data, &ldb, &info);
 
-	return (int) info;
+	return info;
 }
 
 int
@@ -85,14 +85,14 @@ c_linalg_lapack_dgetri (c_matrix *lu, c_vector_int *p)
 	dgetri_(&n, lu->data, &lda, p->data, &wkopt, &lwork, &info);
 
 	lwork = (int) wkopt;
-	if ((int) info != 0 || lwork <= 0) c_error ("c_linalg_lapack_dgetri", "failed to query workspace.");
+	if (info != 0 || lwork <= 0) c_error ("c_linalg_lapack_dgetri", "failed to query workspace.");
 	work = (double *) malloc (lwork * sizeof (double));
 	if (!work) c_error ("c_linalg_lapack_dgetri", "cannot allocate memory work.");
 
 	dgetri_(&n, lu->data, &lda, p->data, work, &lwork, &info);
 	free (work);
 
-	return (int) info;
+	return info;
 }
 
 int
@@ -140,5 +140,5 @@ c_linalg_LU_invert (c_matrix *lu, c_vector_int *p)
 
 	info = c_linalg_lapack_dgetri (lu, p);
 
-	return (int) info;
+	return info;
 }
